Fixed unauthorized image delete/publishing, Unified Tag service, small visual update

+++ PicBook.ApplicationService/ImageService.cs
@@ -20,15 +20,21 @@ namespace PicBook.ApplicationService
         {
             return true;
         }
-        public async Task DeletePic(string id)
+        public async Task DeletePic(string id, string userId)
         {
             var image = await dbimageRepo.FindByIdentifier(id);
-            await dbimageRepo.DeletePic(image);
+            if (image.UserIdentifier.Equals(userId))
+            {
+                await dbimageRepo.DeletePic(image);
+            }
         }
-        public async Task PublicPic(string id)
+        public async Task PublicPic(string id, string userId)
         {
             var image = await dbimageRepo.FindByIdentifier(id);
-            await dbimageRepo.PublicPic(image);
+            if (image.UserIdentifier.Equals(userId))
+            {
+                await dbimageRepo.PublicPic(image);
+            }
         }
         public async Task<Image> UploadImage(byte[] imageBytes, String userIdentifier, String filename)
         {
